npm audit and refactor :black_circle: dependencies out
