+++ package.json
@@ -13,21 +13,21 @@
   },
   "devDependencies": {
     "chai": "3.5.0",
-    "gh-release": "^3.2.1",
+    "gh-release": "^3.4.0",
     "http-server": "^0.10.0",
     "imagemin": "^5.3.1",
     "isparta": "^4.0.0",
     "istanbul": "^0.4.2",
-    "karma": "^2.0.2",
+    "karma": "^3.1.3",
     "karma-chai-sinon": "^0.1.3",
     "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.1",
+    "karma-coverage": "^1.1.2",
     "karma-mocha": "^1.3.0",
     "karma-mocha-reporter": "^2.2.1",
     "karma-sourcemap-loader": "^0.3.5",
     "mkdirp": "^0.5.1",
     "mocha": "^5.2.0",
-    "node-sass": "^3.2.0",
+    "node-sass": "^4.11.0",
     "npm-run-all": "^4.1.2",
     "rollup": "^0.56.5",
     "rollup-plugin-json": "^2.0.0",
