+++ src/helper.js
@@ -1,0 +1,1 @@
+export var WorldGeocodingServiceUrl = 'https://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer/';
